link: Improvements over the runtime API.
This allows the runtime API:

 * To be queried.
 * To be reused across threads.

I need the first to use it on bindgen. The second is not needed right now, but
it's nice to have.

The querying API is really similar to what the `gl` bindings crate uses.

I wonder why a mutex to guard the setting of a thread-local variable, is there
something I missed?