Add Scoop support on windows
[Scoop](https://github.com/lukesampson/scoop) is a popular command-line installer used on Windows with a behavior similar to Homebrew on MacOS. It seems reasonable to have support for the default user installation path.
Instead of fetching the current user and restrict the search to it, I rely on having the glob failing on others users for lack of necessary permissions. If the maintainers feel more comfortable with a specific code path to handle this use case, I can do the extra changes. I also chose the current instead of globbing all the available versions.